refactor(automations): inline workdelivery and split the daemon god file
Mechanical first step of the automations v2 simplification
(docs/plans/2026-07-21-automations-v2-simplification.md):

- Delete internal/workdelivery: the 7-port Ports interface and 51-line
  Service.Deliver pass-through added indirection without depth. Its body
  is inlined verbatim as Daemon.materializeAutomationRun with identical
  error wrapping; the mock-only step-order test is dropped deliberately.
- Rename the former port methods to daemon-local names
  (ensureAutomationTicket, prepareAutomationLocation, ...).
- Split internal/daemon/automations.go (1,587 lines) by concern:
  automations_definitions.go, automations_github.go,
  automations_deliver.go, automations_recover.go; automations.go keeps
  the command handler and shared helpers (176 lines).

No behavior change. Live-verified on the dev profile: manual automation
applied + run reaches delivered through materializeAutomationRun.
